split storage to read and write
