+++ src/main/java/dolla/ui/StorageUi.java
@@ -1,0 +1,5 @@
+package dolla.ui;
+
+public class StorageUi extends Ui {
+
+}
